requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
distro==1.5.0
Django==3.0.7
django-cors-headers==3.3.0
django-sendgrid-v5==0.8.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
matrix-client==0.3.2
mccabe==0.6.1
psycopg2==2.8.5
pycparser==2.20
PyGithub==1.54.1
PyJWT==1.7.1
pylint==2.5.2
pyOpenSSL==19.1.0
python-dotenv==0.13.0
python-http-client==3.2.7
pytz==2020.1
requests==2.24.0
sendgrid==6.3.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
zulip==0.7.0